# Conflicts:
#	pom.xml
#	src/main/java/com/alibaba/excel/converters/string/StringNumberConverter.java
#	src/main/java/com/alibaba/excel/metadata/format/DataFormatter.java
#	src/main/java/com/alibaba/excel/util/NumberDataFormatterUtils.java
#	src/test/java/com/alibaba/easyexcel/test/core/dataformat/DateFormatTest.java
#	src/test/java/com/alibaba/easyexcel/test/temp/Lock2Test.java
#	update.md

+++ src/main/java/com/alibaba/excel/converters/string/StringNumberConverter.java
@@ -51,7 +51,7 @@ public class StringNumberConverter implements Converter<String> {
             && cellData.getDataFormatData().getIndex() != null && !StringUtils.isEmpty(
             cellData.getDataFormatData().getFormat());
         if (hasDataFormatData) {
-            return NumberDataFormatterUtils.format(cellData.getNumberValue().doubleValue(),
+            return NumberDataFormatterUtils.format(cellData.getNumberValue(),
                 cellData.getDataFormatData().getIndex(), cellData.getDataFormatData().getFormat(), globalConfiguration);
         }
         // Default conversion number
